pyproject.toml
[tool.poetry]
name = "lupin"
version = "0.1.0"
description = ""
authors = ["justonelineatatime"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.31.0"
pandas = "^2.2.0"
watchdog = "^3.0.0"
python-chess = "^1.999"
matplotlib = "^3.8.2"
cairosvg = "^2.7.1"
cairocffi = "^1.6.1"
plotly = "^5.18.0"
streamlit-extras = "^0.4.0"
black = "^24.1.1"
google-cloud-bigquery = "^3.17.2"
google-cloud-storage = "^2.14.0"
db-dtypes = "^1.2.0"
streamlit-js-eval = "^0.1.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
